1-1

+++ src/main/java/com/CESI/accessingdatamysql/CompteCourant.java
@@ -15,7 +15,7 @@ public class CompteCourant {
 
     private String Intitule;
     
-    private float Solde;
+    private double Solde;
 
     public Integer getId() {
         return Identifiant;
@@ -43,7 +43,7 @@ public class CompteCourant {
     }
 
 
-    public float get_solde() {
+    public double get_solde() {
         return Solde;
     }
